chore: move AnimationTimer to Game class
